refactor : 클래스 분리
